Move autostop_loop and AUTOSTOP_DELAY to conftest.py
